+++ src/ItemSpawner/Plugin.cs
@@ -13,11 +13,12 @@ using Zorro.Core.CLI;
 namespace ItemSpawner;
 
 [BepInAutoPlugin]
+[ConsoleClassCustomizer("ItemSpawner")]
 public partial class Plugin : BaseUnityPlugin
 {
     internal static ManualLogSource Log { get; private set; } = null!;
     internal static AssetBundle Bundle { get; set; } = null!;
-    internal static KeyCode ToggleKey { get; set; } = KeyCode.F4;
+    internal static KeyCode ToggleKey { get; set; } = KeyCode.F5;
 
     private void Awake()
     {
@@ -25,7 +26,7 @@ public partial class Plugin : BaseUnityPlugin
         string AssetBundlePath = Path.Combine(Path.GetDirectoryName(System.Reflection.Assembly.GetExecutingAssembly().Location), "itemspawnerui");
         Bundle = AssetBundle.LoadFromFile(AssetBundlePath);
 
-        ToggleKey = Config.Bind("General", "ToggleKey", KeyCode.F4, "The button to toggle the item spawner UI.").Value;
+        ToggleKey = Config.Bind("General", "ToggleKey", KeyCode.F5, "The button to toggle the item spawner UI.").Value;
 
         On.GUIManager.Start += On_GUIManager_Start;
 
@@ -45,39 +46,19 @@ public partial class Plugin : BaseUnityPlugin
         }
         GameObject itemSpawnerUI = GameObject.Instantiate(itemSpawnerUIPrefab, self.transform);
         ItemSpawnerWindow spawnerWindow = itemSpawnerUI.AddComponent<ItemSpawnerWindow>();
-
-        InputAction toggleAction;
-        toggleAction = new InputAction(
-            name: "ToggleItemSpawner",
-            type: InputActionType.Button,
-            binding: "<Keyboard>/" + ToggleKey.ToString().ToLower()
-        );
-
-        toggleAction.performed += ctx => spawnerWindow.Toggle();
-        toggleAction.Enable();
+        spawnerWindow.gameObject.SetActive(true);
     }
 
+    [ConsoleCommand]
     public static void Spawn(Item item)
     {
-        if (MainCamera.instance == null)
+        if (!PhotonNetwork.IsConnected || !Character.localCharacter)
         {
             return;
         }
-        if (!PhotonNetwork.IsConnected)
-        {
-            return;
-        }
-        Transform transform = MainCamera.instance.transform;
-        RaycastHit raycastHit;
-        if (!Physics.Raycast(transform.position, transform.forward, out raycastHit, 10f, HelperFunctions.terrainMapMask))
-        {
-            raycastHit = default(RaycastHit);
-        }
-        Vector3 point = raycastHit.point + raycastHit.normal;
 
-        Debug.Log(string.Format("Spawn item: {0} at {1}", item, point));
-        
-        Item spawnedItem = PhotonNetwork.Instantiate("0_Items/" + item.name, point, Quaternion.identity, 0, null).GetComponent<Item>();
-        spawnedItem.view.RPC("RequestPickup", RpcTarget.MasterClient, Character.localCharacter.GetComponent<PhotonView>());
+        Debug.Log(string.Format("Spawn item: {0}", item));
+
+        Character.localCharacter.refs.items.SpawnItemInHand(item.name);
     }
 }